cveda/requirements.txt
numpy==1.26.5
pillow==10.0.1
matplotlib==3.8.2
reportlab==4.1.0
pycocotools==2.0.6
tifffile==2024.11.0
scipy==1.11.3
opencv-python==4.8.1.78
jsonschema==4.21.0
pyyaml==6.0
